Code generator fixes for invokers (#83)

+++ src/Hagar.CodeGenerator/Model/FieldDescription.cs
@@ -4,13 +4,13 @@ namespace Hagar.CodeGenerator
 {
     internal class FieldDescription : IFieldDescription
     {
-        public FieldDescription(uint fieldId, IFieldSymbol field)
+        public FieldDescription(ushort fieldId, IFieldSymbol field)
         {
             FieldId = fieldId;
             Field = field;
         }
         public IFieldSymbol Field { get; }
-        public uint FieldId { get; }
+        public ushort FieldId { get; }
         public ISymbol Member => Field;
         public ITypeSymbol Type => Field.Type;
         public string Name => Field.Name;
