+++ src/Hagar.CodeGenerator/Model/IMemberDescription.cs
@@ -4,7 +4,7 @@ namespace Hagar.CodeGenerator
 {
     internal interface IMemberDescription
     {
-        uint FieldId { get; }
+        ushort FieldId { get; }
         ISymbol Member { get; }
         ITypeSymbol Type { get; }
         string Name { get; }
